relocate functions in module
